% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddBandingRequest}
\alias{AddBandingRequest}
\title{AddBandingRequest Object}
\usage{
AddBandingRequest(bandedRange = NULL)
}
\arguments{
\item{bandedRange}{The banded range to add}
}
\value{
AddBandingRequest object
}
\description{
AddBandingRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds a new banded range to the spreadsheet.
}
