% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddBandingResponse}
\alias{AddBandingResponse}
\title{AddBandingResponse Object}
\usage{
AddBandingResponse(bandedRange = NULL)
}
\arguments{
\item{bandedRange}{The banded range that was added}
}
\value{
AddBandingResponse object
}
\description{
AddBandingResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a banded range.
}
