% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddChartRequest}
\alias{AddChartRequest}
\title{AddChartRequest Object}
\usage{
AddChartRequest(chart = NULL)
}
\arguments{
\item{chart}{The chart that should be added to the spreadsheet, including the position}
}
\value{
AddChartRequest object
}
\description{
AddChartRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds a chart to a sheet in the spreadsheet.
}
