% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddChartResponse}
\alias{AddChartResponse}
\title{AddChartResponse Object}
\usage{
AddChartResponse(chart = NULL)
}
\arguments{
\item{chart}{The newly added chart}
}
\value{
AddChartResponse object
}
\description{
AddChartResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a chart to a spreadsheet.
}
