% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddConditionalFormatRuleRequest}
\alias{AddConditionalFormatRuleRequest}
\title{AddConditionalFormatRuleRequest Object}
\usage{
AddConditionalFormatRuleRequest(rule = NULL, index = NULL)
}
\arguments{
\item{rule}{The rule to add}

\item{index}{The zero-based index where the rule should be inserted}
}
\value{
AddConditionalFormatRuleRequest object
}
\description{
AddConditionalFormatRuleRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds a new conditional format rule at the given index.All subsequent rules' indexes are incremented.
}
