% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddFilterViewRequest}
\alias{AddFilterViewRequest}
\title{AddFilterViewRequest Object}
\usage{
AddFilterViewRequest(filter = NULL)
}
\arguments{
\item{filter}{The filter to add}
}
\value{
AddFilterViewRequest object
}
\description{
AddFilterViewRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds a filter view.
}
