% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddFilterViewResponse}
\alias{AddFilterViewResponse}
\title{AddFilterViewResponse Object}
\usage{
AddFilterViewResponse(filter = NULL)
}
\arguments{
\item{filter}{The newly added filter view}
}
\value{
AddFilterViewResponse object
}
\description{
AddFilterViewResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a filter view.
}
