% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddNamedRangeRequest}
\alias{AddNamedRangeRequest}
\title{AddNamedRangeRequest Object}
\usage{
AddNamedRangeRequest(namedRange = NULL)
}
\arguments{
\item{namedRange}{The named range to add}
}
\value{
AddNamedRangeRequest object
}
\description{
AddNamedRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds a named range to the spreadsheet.
}
