% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddNamedRangeResponse}
\alias{AddNamedRangeResponse}
\title{Google Sheets API Objects 
Reads and writes Google Sheets.}
\usage{
AddNamedRangeResponse(namedRange = NULL)
}
\arguments{
\item{namedRange}{The named range to add}
}
\value{
AddNamedRangeResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-04 00:02:25
filename: /Users/mark/dev/R/autoGoogleAPI/googlesheetsv4.auto/R/sheets_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AddNamedRangeResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a named range.
}

