% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddProtectedRangeRequest}
\alias{AddProtectedRangeRequest}
\title{AddProtectedRangeRequest Object}
\usage{
AddProtectedRangeRequest(protectedRange = NULL)
}
\arguments{
\item{protectedRange}{The protected range to be added}
}
\value{
AddProtectedRangeRequest object
}
\description{
AddProtectedRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds a new protected range.
}
