% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddProtectedRangeResponse}
\alias{AddProtectedRangeResponse}
\title{AddProtectedRangeResponse Object}
\usage{
AddProtectedRangeResponse(protectedRange = NULL)
}
\arguments{
\item{protectedRange}{The newly added protected range}
}
\value{
AddProtectedRangeResponse object
}
\description{
AddProtectedRangeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a new protected range.
}
