% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddSheetRequest}
\alias{AddSheetRequest}
\title{AddSheetRequest Object}
\usage{
AddSheetRequest(properties = NULL)
}
\arguments{
\item{properties}{The properties the new sheet should have}
}
\value{
AddSheetRequest object
}
\description{
AddSheetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Adds a new sheet.When a sheet is added at a given index,all subsequent sheets' indexes are incremented.To add an object sheet, use AddChartRequest instead and specifyEmbeddedObjectPosition.sheetId orEmbeddedObjectPosition.newSheet.
}
