% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddSheetResponse}
\alias{AddSheetResponse}
\title{AddSheetResponse Object}
\usage{
AddSheetResponse(properties = NULL)
}
\arguments{
\item{properties}{The properties of the newly added sheet}
}
\value{
AddSheetResponse object
}
\description{
AddSheetResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a sheet.
}
