% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AppendDimensionRequest}
\alias{AppendDimensionRequest}
\title{Google Sheets API Objects 
Reads and writes Google Sheets.}
\usage{
AppendDimensionRequest(dimension = NULL, length = NULL, sheetId = NULL)
}
\arguments{
\item{dimension}{Whether rows or columns should be appended}

\item{length}{The number of rows or columns to append}

\item{sheetId}{The sheet to append rows or columns to}
}
\value{
AppendDimensionRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:13:27
filename: /Users/mark/dev/R/autoGoogleAPI/googlesheetsv4.auto/R/sheets_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AppendDimensionRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Appends rows or columns to the end of a sheet.
}
