% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AppendValuesResponse}
\alias{AppendValuesResponse}
\title{AppendValuesResponse Object}
\usage{
AppendValuesResponse(spreadsheetId = NULL, updates = NULL,
  tableRange = NULL)
}
\arguments{
\item{spreadsheetId}{The spreadsheet the updates were applied to}

\item{updates}{Information about the updates that were applied}

\item{tableRange}{The range (in A1 notation) of the table that values are being appended to}
}
\value{
AppendValuesResponse object
}
\description{
AppendValuesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when updating a range of values in a spreadsheet.
}
