% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AutoFillRequest}
\alias{AutoFillRequest}
\title{AutoFillRequest Object}
\usage{
AutoFillRequest(useAlternateSeries = NULL, range = NULL,
  sourceAndDestination = NULL)
}
\arguments{
\item{useAlternateSeries}{True if we should generate data with the 'alternate' series}

\item{range}{The range to autofill}

\item{sourceAndDestination}{The source and destination areas to autofill}
}
\value{
AutoFillRequest object
}
\description{
AutoFillRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Fills in more data based on existing data.
}

