% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AutoResizeDimensionsRequest}
\alias{AutoResizeDimensionsRequest}
\title{AutoResizeDimensionsRequest Object}
\usage{
AutoResizeDimensionsRequest(dimensions = NULL)
}
\arguments{
\item{dimensions}{The dimensions to automatically resize}
}
\value{
AutoResizeDimensionsRequest object
}
\description{
AutoResizeDimensionsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Automatically resizes one or more dimensions based on the contentsof the cells in that dimension.
}
