% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BandedRange}
\alias{BandedRange}
\title{BandedRange Object}
\usage{
BandedRange(rowProperties = NULL, columnProperties = NULL, range = NULL,
  bandedRangeId = NULL)
}
\arguments{
\item{rowProperties}{Properties for row bands}

\item{columnProperties}{Properties for column bands}

\item{range}{The range over which these properties are applied}

\item{bandedRangeId}{The id of the banded range}
}
\value{
BandedRange object
}
\description{
BandedRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A banded (alternating colors) range in a sheet.
}
