% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BandingProperties}
\alias{BandingProperties}
\title{BandingProperties Object}
\usage{
BandingProperties(footerColor = NULL, headerColor = NULL,
  firstBandColor = NULL, secondBandColor = NULL)
}
\arguments{
\item{footerColor}{The color of the last row or column}

\item{headerColor}{The color of the first row or column}

\item{firstBandColor}{The first color that is alternating}

\item{secondBandColor}{The second color that is alternating}
}
\value{
BandingProperties object
}
\description{
BandingProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties referring a single dimension (either row or column). If bothBandedRange.row_properties and BandedRange.column_properties areset, the fill colors are applied to cells according to the following rules:* header_color and footer_color take priority over band colors.* first_band_color takes priority over second_band_color.* row_properties takes priority over column_properties.For example, the first row color takes priority over the first columncolor, but the first column color takes priority over the second row color.Similarly, the row header takes priority over the column header in thetop left cell, but the column header takes priority over the first rowcolor if the row header is not set.
}
