% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BasicChartAxis}
\alias{BasicChartAxis}
\title{BasicChartAxis Object}
\usage{
BasicChartAxis(position = NULL, title = NULL, format = NULL)
}
\arguments{
\item{position}{The position of this axis}

\item{title}{The title of this axis}

\item{format}{The format of the title}
}
\value{
BasicChartAxis object
}
\description{
BasicChartAxis Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An axis of the chart.A chart may not have more than one axis peraxis position.
}

