% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BasicChartDomain}
\alias{BasicChartDomain}
\title{BasicChartDomain Object}
\usage{
BasicChartDomain(domain = NULL)
}
\arguments{
\item{domain}{The data of the domain}
}
\value{
BasicChartDomain object
}
\description{
BasicChartDomain Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The domain of a chart.For example, if charting stock prices over time, this would be the date.
}

