% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BasicChartSeries}
\alias{BasicChartSeries}
\title{BasicChartSeries Object}
\usage{
BasicChartSeries(series = NULL, type = NULL, targetAxis = NULL)
}
\arguments{
\item{series}{The data being visualized in this chart series}

\item{type}{The type of this series}

\item{targetAxis}{The minor axis that will specify the range of values for this series}
}
\value{
BasicChartSeries object
}
\description{
BasicChartSeries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single series of data in a chart.For example, if charting stock prices over time, multiple series may exist,one for the 'Open Price', 'High Price', 'Low Price' and 'Close Price'.
}
