% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BasicFilter.criteria}
\alias{BasicFilter.criteria}
\title{BasicFilter.criteria Object}
\usage{
BasicFilter.criteria()
}
\value{
BasicFilter.criteria object
}
\description{
BasicFilter.criteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The criteria for showing/hiding values per column.The map's key is the column index, and the value is the criteria forthat column.
}
\seealso{
Other BasicFilter functions: \code{\link{BasicFilter}}
}
