% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BasicFilter}
\alias{BasicFilter}
\title{BasicFilter Object}
\usage{
BasicFilter(BasicFilter.criteria = NULL, range = NULL, criteria = NULL,
  sortSpecs = NULL)
}
\arguments{
\item{BasicFilter.criteria}{The \link{BasicFilter.criteria} object or list of objects}

\item{range}{The range the filter covers}

\item{criteria}{The criteria for showing/hiding values per column}

\item{sortSpecs}{The sort order per column}
}
\value{
BasicFilter object
}
\description{
BasicFilter Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The default filter associated with a sheet.
}
\seealso{
Other BasicFilter functions: \code{\link{BasicFilter.criteria}}
}
