% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchClearValuesRequest}
\alias{BatchClearValuesRequest}
\title{BatchClearValuesRequest Object}
\usage{
BatchClearValuesRequest(ranges = NULL)
}
\arguments{
\item{ranges}{The ranges to clear, in A1 notation}
}
\value{
BatchClearValuesRequest object
}
\description{
BatchClearValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for clearing more than one range of values in a spreadsheet.
}
