% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchUpdateSpreadsheetResponse}
\alias{BatchUpdateSpreadsheetResponse}
\title{BatchUpdateSpreadsheetResponse Object}
\usage{
BatchUpdateSpreadsheetResponse(replies = NULL, updatedSpreadsheet = NULL,
  spreadsheetId = NULL)
}
\arguments{
\item{replies}{The reply of the updates}

\item{updatedSpreadsheet}{The spreadsheet after updates were applied}

\item{spreadsheetId}{The spreadsheet the updates were applied to}
}
\value{
BatchUpdateSpreadsheetResponse object
}
\description{
BatchUpdateSpreadsheetResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The reply for batch updating a spreadsheet.
}
