% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchUpdateValuesRequest}
\alias{BatchUpdateValuesRequest}
\title{BatchUpdateValuesRequest Object}
\usage{
BatchUpdateValuesRequest(valueInputOption = NULL, data = NULL)
}
\arguments{
\item{valueInputOption}{How the input data should be interpreted}

\item{data}{The new values to apply to the spreadsheet}
}
\value{
BatchUpdateValuesRequest object
}
\description{
BatchUpdateValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for updating more than one range of values in a spreadsheet.
}

