% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BatchUpdateValuesResponse}
\alias{BatchUpdateValuesResponse}
\title{BatchUpdateValuesResponse Object}
\usage{
BatchUpdateValuesResponse(responses = NULL, totalUpdatedSheets = NULL,
  totalUpdatedCells = NULL, totalUpdatedColumns = NULL,
  spreadsheetId = NULL, totalUpdatedRows = NULL)
}
\arguments{
\item{responses}{One UpdateValuesResponse per requested range, in the same order as}

\item{totalUpdatedSheets}{The total number of sheets where at least one cell in the sheet was}

\item{totalUpdatedCells}{The total number of cells updated}

\item{totalUpdatedColumns}{The total number of columns where at least one cell in the column was}

\item{spreadsheetId}{The spreadsheet the updates were applied to}

\item{totalUpdatedRows}{The total number of rows where at least one cell in the row was updated}
}
\value{
BatchUpdateValuesResponse object
}
\description{
BatchUpdateValuesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when updating a range of values in a spreadsheet.
}
