% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BooleanCondition}
\alias{BooleanCondition}
\title{BooleanCondition Object}
\usage{
BooleanCondition(values = NULL, type = NULL)
}
\arguments{
\item{values}{The values of the condition}

\item{type}{The type of condition}
}
\value{
BooleanCondition object
}
\description{
BooleanCondition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A condition that can evaluate to true or false.BooleanConditions are used by conditional formatting,data validation, and the criteria in filters.
}

