% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Borders}
\alias{Borders}
\title{Borders Object}
\usage{
Borders(right = NULL, bottom = NULL, top = NULL, left = NULL)
}
\arguments{
\item{right}{The right border of the cell}

\item{bottom}{The bottom border of the cell}

\item{top}{The top border of the cell}

\item{left}{The left border of the cell}
}
\value{
Borders object
}
\description{
Borders Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The borders of the cell.
}
