% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ChartData}
\alias{ChartData}
\title{ChartData Object}
\usage{
ChartData(sourceRange = NULL)
}
\arguments{
\item{sourceRange}{The source ranges of the data}
}
\value{
ChartData object
}
\description{
ChartData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The data included in a domain or series.
}
