% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ChartSourceRange}
\alias{ChartSourceRange}
\title{ChartSourceRange Object}
\usage{
ChartSourceRange(sources = NULL)
}
\arguments{
\item{sources}{The ranges of data for a series or domain}
}
\value{
ChartSourceRange object
}
\description{
ChartSourceRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Source ranges for a chart.
}
