% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ChartSpec}
\alias{ChartSpec}
\title{ChartSpec Object}
\usage{
ChartSpec(hiddenDimensionStrategy = NULL, basicChart = NULL, title = NULL,
  pieChart = NULL)
}
\arguments{
\item{hiddenDimensionStrategy}{Determines how the charts will use hidden rows or columns}

\item{basicChart}{A basic chart specification, can be one of many kinds of charts}

\item{title}{The title of the chart}

\item{pieChart}{A pie chart specification}
}
\value{
ChartSpec object
}
\description{
ChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The specifications of a chart.
}

