% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ClearBasicFilterRequest}
\alias{ClearBasicFilterRequest}
\title{ClearBasicFilterRequest Object}
\usage{
ClearBasicFilterRequest(sheetId = NULL)
}
\arguments{
\item{sheetId}{The sheet ID on which the basic filter should be cleared}
}
\value{
ClearBasicFilterRequest object
}
\description{
ClearBasicFilterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Clears the basic filter, if any exists on the sheet.
}
