% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ClearValuesRequest}
\alias{ClearValuesRequest}
\title{ClearValuesRequest Object}
\usage{
ClearValuesRequest()
}
\value{
ClearValuesRequest object
}
\description{
ClearValuesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request for clearing a range of values in a spreadsheet.
}
