% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ClearValuesResponse}
\alias{ClearValuesResponse}
\title{ClearValuesResponse Object}
\usage{
ClearValuesResponse(spreadsheetId = NULL, clearedRange = NULL)
}
\arguments{
\item{spreadsheetId}{The spreadsheet the updates were applied to}

\item{clearedRange}{The range (in A1 notation) that was cleared}
}
\value{
ClearValuesResponse object
}
\description{
ClearValuesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when clearing a range of values in a spreadsheet.
}
