% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ConditionalFormatRule}
\alias{ConditionalFormatRule}
\title{ConditionalFormatRule Object}
\usage{
ConditionalFormatRule(ranges = NULL, gradientRule = NULL,
  booleanRule = NULL)
}
\arguments{
\item{ranges}{The ranges that will be formatted if the condition is true}

\item{gradientRule}{The formatting will vary based on the gradients in the rule}

\item{booleanRule}{The formatting is either 'on' or 'off' according to the rule}
}
\value{
ConditionalFormatRule object
}
\description{
ConditionalFormatRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rule describing a conditional format.
}
