% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ConditionValue}
\alias{ConditionValue}
\title{ConditionValue Object}
\usage{
ConditionValue(relativeDate = NULL, userEnteredValue = NULL)
}
\arguments{
\item{relativeDate}{A relative date (based on the current date)}

\item{userEnteredValue}{A value the condition is based on}
}
\value{
ConditionValue object
}
\description{
ConditionValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The value of the condition.
}
