% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CopyPasteRequest}
\alias{CopyPasteRequest}
\title{CopyPasteRequest Object}
\usage{
CopyPasteRequest(pasteType = NULL, pasteOrientation = NULL, source = NULL,
  destination = NULL)
}
\arguments{
\item{pasteType}{What kind of data to paste}

\item{pasteOrientation}{How that data should be oriented when pasting}

\item{source}{The source range to copy}

\item{destination}{The location to paste to}
}
\value{
CopyPasteRequest object
}
\description{
CopyPasteRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Copies data from the source to the destination.
}

