% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CopySheetToAnotherSpreadsheetRequest}
\alias{CopySheetToAnotherSpreadsheetRequest}
\title{CopySheetToAnotherSpreadsheetRequest Object}
\usage{
CopySheetToAnotherSpreadsheetRequest(destinationSpreadsheetId = NULL)
}
\arguments{
\item{destinationSpreadsheetId}{The ID of the spreadsheet to copy the sheet to}
}
\value{
CopySheetToAnotherSpreadsheetRequest object
}
\description{
CopySheetToAnotherSpreadsheetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request to copy a sheet across spreadsheets.
}
