% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteBandingRequest}
\alias{DeleteBandingRequest}
\title{DeleteBandingRequest Object}
\usage{
DeleteBandingRequest(bandedRangeId = NULL)
}
\arguments{
\item{bandedRangeId}{The ID of the banded range to delete}
}
\value{
DeleteBandingRequest object
}
\description{
DeleteBandingRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Removes the banded range with the given ID from the spreadsheet.
}
