% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteConditionalFormatRuleResponse}
\alias{DeleteConditionalFormatRuleResponse}
\title{DeleteConditionalFormatRuleResponse Object}
\usage{
DeleteConditionalFormatRuleResponse(rule = NULL)
}
\arguments{
\item{rule}{The rule that was deleted}
}
\value{
DeleteConditionalFormatRuleResponse object
}
\description{
DeleteConditionalFormatRuleResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of deleting a conditional format rule.
}
