% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteDimensionRequest}
\alias{DeleteDimensionRequest}
\title{DeleteDimensionRequest Object}
\usage{
DeleteDimensionRequest(range = NULL)
}
\arguments{
\item{range}{The dimensions to delete from the sheet}
}
\value{
DeleteDimensionRequest object
}
\description{
DeleteDimensionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes the dimensions from the sheet.
}
