% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteEmbeddedObjectRequest}
\alias{DeleteEmbeddedObjectRequest}
\title{DeleteEmbeddedObjectRequest Object}
\usage{
DeleteEmbeddedObjectRequest(objectId = NULL)
}
\arguments{
\item{objectId}{The ID of the embedded object to delete}
}
\value{
DeleteEmbeddedObjectRequest object
}
\description{
DeleteEmbeddedObjectRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes the embedded object with the given ID.
}
