% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteFilterViewRequest}
\alias{DeleteFilterViewRequest}
\title{DeleteFilterViewRequest Object}
\usage{
DeleteFilterViewRequest(filterId = NULL)
}
\arguments{
\item{filterId}{The ID of the filter to delete}
}
\value{
DeleteFilterViewRequest object
}
\description{
DeleteFilterViewRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes a particular filter view.
}
