% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteNamedRangeRequest}
\alias{DeleteNamedRangeRequest}
\title{DeleteNamedRangeRequest Object}
\usage{
DeleteNamedRangeRequest(namedRangeId = NULL)
}
\arguments{
\item{namedRangeId}{The ID of the named range to delete}
}
\value{
DeleteNamedRangeRequest object
}
\description{
DeleteNamedRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Removes the named range with the given ID from the spreadsheet.
}
