% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteProtectedRangeRequest}
\alias{DeleteProtectedRangeRequest}
\title{DeleteProtectedRangeRequest Object}
\usage{
DeleteProtectedRangeRequest(protectedRangeId = NULL)
}
\arguments{
\item{protectedRangeId}{The ID of the protected range to delete}
}
\value{
DeleteProtectedRangeRequest object
}
\description{
DeleteProtectedRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes the protected range with the given ID.
}
