% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteRangeRequest}
\alias{DeleteRangeRequest}
\title{DeleteRangeRequest Object}
\usage{
DeleteRangeRequest(range = NULL, shiftDimension = NULL)
}
\arguments{
\item{range}{The range of cells to delete}

\item{shiftDimension}{The dimension from which deleted cells will be replaced with}
}
\value{
DeleteRangeRequest object
}
\description{
DeleteRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes a range of cells, shifting other cells into the deleted area.
}
