% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteSheetRequest}
\alias{DeleteSheetRequest}
\title{DeleteSheetRequest Object}
\usage{
DeleteSheetRequest(sheetId = NULL)
}
\arguments{
\item{sheetId}{The ID of the sheet to delete}
}
\value{
DeleteSheetRequest object
}
\description{
DeleteSheetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes the requested sheet.
}
