% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DimensionProperties}
\alias{DimensionProperties}
\title{DimensionProperties Object}
\usage{
DimensionProperties(pixelSize = NULL, hiddenByFilter = NULL,
  hiddenByUser = NULL)
}
\arguments{
\item{pixelSize}{The height (if a row) or width (if a column) of the dimension in pixels}

\item{hiddenByFilter}{True if this dimension is being filtered}

\item{hiddenByUser}{True if this dimension is explicitly hidden}
}
\value{
DimensionProperties object
}
\description{
DimensionProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties about a dimension.
}
