% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DuplicateFilterViewRequest}
\alias{DuplicateFilterViewRequest}
\title{DuplicateFilterViewRequest Object}
\usage{
DuplicateFilterViewRequest(filterId = NULL)
}
\arguments{
\item{filterId}{The ID of the filter being duplicated}
}
\value{
DuplicateFilterViewRequest object
}
\description{
DuplicateFilterViewRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Duplicates a particular filter view.
}
