% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DuplicateFilterViewResponse}
\alias{DuplicateFilterViewResponse}
\title{DuplicateFilterViewResponse Object}
\usage{
DuplicateFilterViewResponse(filter = NULL)
}
\arguments{
\item{filter}{The newly created filter}
}
\value{
DuplicateFilterViewResponse object
}
\description{
DuplicateFilterViewResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of a filter view being duplicated.
}
