% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DuplicateSheetRequest}
\alias{DuplicateSheetRequest}
\title{DuplicateSheetRequest Object}
\usage{
DuplicateSheetRequest(newSheetId = NULL, insertSheetIndex = NULL,
  newSheetName = NULL, sourceSheetId = NULL)
}
\arguments{
\item{newSheetId}{If set, the ID of the new sheet}

\item{insertSheetIndex}{The zero-based index where the new sheet should be inserted}

\item{newSheetName}{The name of the new sheet}

\item{sourceSheetId}{The sheet to duplicate}
}
\value{
DuplicateSheetRequest object
}
\description{
DuplicateSheetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Duplicates the contents of a sheet.
}
