% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DuplicateSheetResponse}
\alias{DuplicateSheetResponse}
\title{DuplicateSheetResponse Object}
\usage{
DuplicateSheetResponse(properties = NULL)
}
\arguments{
\item{properties}{The properties of the duplicate sheet}
}
\value{
DuplicateSheetResponse object
}
\description{
DuplicateSheetResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of duplicating a sheet.
}
