% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ErrorValue}
\alias{ErrorValue}
\title{ErrorValue Object}
\usage{
ErrorValue(type = NULL, message = NULL)
}
\arguments{
\item{type}{The type of error}

\item{message}{A message with more information about the error}
}
\value{
ErrorValue object
}
\description{
ErrorValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An error in a cell.
}
