% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ExtendedValue}
\alias{ExtendedValue}
\title{ExtendedValue Object}
\usage{
ExtendedValue(formulaValue = NULL, errorValue = NULL, boolValue = NULL,
  numberValue = NULL, stringValue = NULL)
}
\arguments{
\item{formulaValue}{Represents a formula}

\item{errorValue}{Represents an error}

\item{boolValue}{Represents a boolean value}

\item{numberValue}{Represents a double value}

\item{stringValue}{Represents a string value}
}
\value{
ExtendedValue object
}
\description{
ExtendedValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The kinds of value that a cell in a spreadsheet can have.
}

