% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{FilterView.criteria}
\alias{FilterView.criteria}
\title{FilterView.criteria Object}
\usage{
FilterView.criteria()
}
\value{
FilterView.criteria object
}
\description{
FilterView.criteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The criteria for showing/hiding values per column.The map's key is the column index, and the value is the criteria forthat column.
}
\seealso{
Other FilterView functions: \code{\link{FilterView}}
}
