% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{FilterView}
\alias{FilterView}
\title{FilterView Object}
\usage{
FilterView(FilterView.criteria = NULL, title = NULL, namedRangeId = NULL,
  sortSpecs = NULL, range = NULL, criteria = NULL, filterViewId = NULL)
}
\arguments{
\item{FilterView.criteria}{The \link{FilterView.criteria} object or list of objects}

\item{title}{The name of the filter view}

\item{namedRangeId}{The named range this filter view is backed by, if any}

\item{sortSpecs}{The sort order per column}

\item{range}{The range this filter view covers}

\item{criteria}{The criteria for showing/hiding values per column}

\item{filterViewId}{The ID of the filter view}
}
\value{
FilterView object
}
\description{
FilterView Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A filter view.
}
\seealso{
Other FilterView functions: \code{\link{FilterView.criteria}}
}

