% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{GridData}
\alias{GridData}
\title{GridData Object}
\usage{
GridData(columnMetadata = NULL, rowData = NULL, startRow = NULL,
  rowMetadata = NULL, startColumn = NULL)
}
\arguments{
\item{columnMetadata}{Metadata about the requested columns in the grid, starting with the column}

\item{rowData}{The data in the grid, one entry per row,}

\item{startRow}{The first row this GridData refers to, zero-based}

\item{rowMetadata}{Metadata about the requested rows in the grid, starting with the row}

\item{startColumn}{The first column this GridData refers to, zero-based}
}
\value{
GridData object
}
\description{
GridData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data in the grid, as well as metadata about the dimensions.
}

