% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{GridProperties}
\alias{GridProperties}
\title{GridProperties Object}
\usage{
GridProperties(rowCount = NULL, frozenRowCount = NULL,
  hideGridlines = NULL, columnCount = NULL, frozenColumnCount = NULL)
}
\arguments{
\item{rowCount}{The number of rows in the grid}

\item{frozenRowCount}{The number of rows that are frozen in the grid}

\item{hideGridlines}{True if the grid isn't showing gridlines in the UI}

\item{columnCount}{The number of columns in the grid}

\item{frozenColumnCount}{The number of columns that are frozen in the grid}
}
\value{
GridProperties object
}
\description{
GridProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of a grid.
}
