% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{GridRange}
\alias{GridRange}
\title{GridRange Object}
\usage{
GridRange(startColumnIndex = NULL, sheetId = NULL, endRowIndex = NULL,
  endColumnIndex = NULL, startRowIndex = NULL)
}
\arguments{
\item{startColumnIndex}{The start column (inclusive) of the range, or not set if unbounded}

\item{sheetId}{The sheet this range is on}

\item{endRowIndex}{The end row (exclusive) of the range, or not set if unbounded}

\item{endColumnIndex}{The end column (exclusive) of the range, or not set if unbounded}

\item{startRowIndex}{The start row (inclusive) of the range, or not set if unbounded}
}
\value{
GridRange object
}
\description{
GridRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A range on a sheet.All indexes are zero-based.Indexes are half open, e.g the start index is inclusiveand the end index is exclusive -- [start_index, end_index).Missing indexes indicate the range is unbounded on that side.For example, if `'Sheet1'` is sheet ID 0, then:  `Sheet1!A1:A1 == sheet_id: 0,                  start_row_index: 0, end_row_index: 1,                  start_column_index: 0, end_column_index: 1`  `Sheet1!A3:B4 == sheet_id: 0,                  start_row_index: 2, end_row_index: 4,                  start_column_index: 0, end_column_index: 2`  `Sheet1!A:B == sheet_id: 0,                start_column_index: 0, end_column_index: 2`  `Sheet1!A5:B == sheet_id: 0,                 start_row_index: 4,                 start_column_index: 0, end_column_index: 2`  `Sheet1 == sheet_id:0`The start index must always be less than or equal to the end index.If the start index equals the end index, then the range is empty.Empty ranges are typically not meaningful and are usually rendered in theUI as `#REF!`.
}
