% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{InsertDimensionRequest}
\alias{InsertDimensionRequest}
\title{InsertDimensionRequest Object}
\usage{
InsertDimensionRequest(inheritFromBefore = NULL, range = NULL)
}
\arguments{
\item{inheritFromBefore}{Whether dimension properties should be extended from the dimensions}

\item{range}{The dimensions to insert}
}
\value{
InsertDimensionRequest object
}
\description{
InsertDimensionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Inserts rows or columns in a sheet at a particular index.
}
