% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{InsertRangeRequest}
\alias{InsertRangeRequest}
\title{InsertRangeRequest Object}
\usage{
InsertRangeRequest(shiftDimension = NULL, range = NULL)
}
\arguments{
\item{shiftDimension}{The dimension which will be shifted when inserting cells}

\item{range}{The range to insert new cells into}
}
\value{
InsertRangeRequest object
}
\description{
InsertRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Inserts cells into a range, shifting the existing cells over or down.
}
