% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{IterativeCalculationSettings}
\alias{IterativeCalculationSettings}
\title{IterativeCalculationSettings Object}
\usage{
IterativeCalculationSettings(convergenceThreshold = NULL,
  maxIterations = NULL)
}
\arguments{
\item{convergenceThreshold}{When iterative calculation is enabled, the threshold value such that}

\item{maxIterations}{When iterative calculation is enabled, the maximum number of calculation}
}
\value{
IterativeCalculationSettings object
}
\description{
IterativeCalculationSettings Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Settings to control how circular dependencies are resolved with iterativecalculation.
}
