% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{MergeCellsRequest}
\alias{MergeCellsRequest}
\title{MergeCellsRequest Object}
\usage{
MergeCellsRequest(mergeType = NULL, range = NULL)
}
\arguments{
\item{mergeType}{How the cells should be merged}

\item{range}{The range of cells to merge}
}
\value{
MergeCellsRequest object
}
\description{
MergeCellsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Merges all cells in the range.
}
