% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{MoveDimensionRequest}
\alias{MoveDimensionRequest}
\title{MoveDimensionRequest Object}
\usage{
MoveDimensionRequest(destinationIndex = NULL, source = NULL)
}
\arguments{
\item{destinationIndex}{The zero-based start index of where to move the source data to,}

\item{source}{The source dimensions to move}
}
\value{
MoveDimensionRequest object
}
\description{
MoveDimensionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Moves one or more rows or columns.
}
