% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{NamedRange}
\alias{NamedRange}
\title{NamedRange Object}
\usage{
NamedRange(namedRangeId = NULL, range = NULL, name = NULL)
}
\arguments{
\item{namedRangeId}{The ID of the named range}

\item{range}{The range this represents}

\item{name}{The name of the named range}
}
\value{
NamedRange object
}
\description{
NamedRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A named range.
}

