% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{NumberFormat}
\alias{NumberFormat}
\title{NumberFormat Object}
\usage{
NumberFormat(pattern = NULL, type = NULL)
}
\arguments{
\item{pattern}{Pattern string used for formatting}

\item{type}{The type of the number format}
}
\value{
NumberFormat object
}
\description{
NumberFormat Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The number format of a cell.
}
