% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{OverlayPosition}
\alias{OverlayPosition}
\title{OverlayPosition Object}
\usage{
OverlayPosition(offsetXPixels = NULL, anchorCell = NULL,
  offsetYPixels = NULL, heightPixels = NULL, widthPixels = NULL)
}
\arguments{
\item{offsetXPixels}{The horizontal offset, in pixels, that the object is offset}

\item{anchorCell}{The cell the object is anchored to}

\item{offsetYPixels}{The vertical offset, in pixels, that the object is offset}

\item{heightPixels}{The height of the object, in pixels}

\item{widthPixels}{The width of the object, in pixels}
}
\value{
OverlayPosition object
}
\description{
OverlayPosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The location an object is overlaid on top of a grid.
}
