% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Padding}
\alias{Padding}
\title{Padding Object}
\usage{
Padding(right = NULL, left = NULL, top = NULL, bottom = NULL)
}
\arguments{
\item{right}{The right padding of the cell}

\item{left}{The left padding of the cell}

\item{top}{The top padding of the cell}

\item{bottom}{The bottom padding of the cell}
}
\value{
Padding object
}
\description{
Padding Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The amount of padding around the cell, in pixels.When updating padding, every field must be specified.
}

