% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PasteDataRequest}
\alias{PasteDataRequest}
\title{PasteDataRequest Object}
\usage{
PasteDataRequest(data = NULL, coordinate = NULL, delimiter = NULL,
  type = NULL, html = NULL)
}
\arguments{
\item{data}{The data to insert}

\item{coordinate}{The coordinate at which the data should start being inserted}

\item{delimiter}{The delimiter in the data}

\item{type}{How the data should be pasted}

\item{html}{True if the data is HTML}
}
\value{
PasteDataRequest object
}
\description{
PasteDataRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Inserts data into the spreadsheet starting at the specified coordinate.
}

