% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PieChartSpec}
\alias{PieChartSpec}
\title{PieChartSpec Object}
\usage{
PieChartSpec(series = NULL, legendPosition = NULL, pieHole = NULL,
  domain = NULL, threeDimensional = NULL)
}
\arguments{
\item{series}{The data that covers the one and only series of the pie chart}

\item{legendPosition}{Where the legend of the pie chart should be drawn}

\item{pieHole}{The size of the hole in the pie chart}

\item{domain}{The data that covers the domain of the pie chart}

\item{threeDimensional}{True if the pie is three dimensional}
}
\value{
PieChartSpec object
}
\description{
PieChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A <a href='/chart/interactive/docs/gallery/piechart'>pie chart</a>.
}
