% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotFilterCriteria}
\alias{PivotFilterCriteria}
\title{PivotFilterCriteria Object}
\usage{
PivotFilterCriteria(visibleValues = NULL)
}
\arguments{
\item{visibleValues}{Values that should be included}
}
\value{
PivotFilterCriteria object
}
\description{
PivotFilterCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Criteria for showing/hiding rows in a pivot table.
}
