% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotGroup}
\alias{PivotGroup}
\title{PivotGroup Object}
\usage{
PivotGroup(sortOrder = NULL, sourceColumnOffset = NULL, showTotals = NULL,
  valueBucket = NULL, valueMetadata = NULL)
}
\arguments{
\item{sortOrder}{The order the values in this group should be sorted}

\item{sourceColumnOffset}{The column offset of the source range that this grouping is based on}

\item{showTotals}{True if the pivot table should include the totals for this grouping}

\item{valueBucket}{The bucket of the opposite pivot group to sort by}

\item{valueMetadata}{Metadata about values in the grouping}
}
\value{
PivotGroup object
}
\description{
PivotGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single grouping (either row or column) in a pivot table.
}

