% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotGroupValueMetadata}
\alias{PivotGroupValueMetadata}
\title{PivotGroupValueMetadata Object}
\usage{
PivotGroupValueMetadata(value = NULL, collapsed = NULL)
}
\arguments{
\item{value}{The calculated value the metadata corresponds to}

\item{collapsed}{True if the data corresponding to the value is collapsed}
}
\value{
PivotGroupValueMetadata object
}
\description{
PivotGroupValueMetadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata about a value in a pivot grouping.
}
