% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotTable.criteria}
\alias{PivotTable.criteria}
\title{PivotTable.criteria Object}
\usage{
PivotTable.criteria()
}
\value{
PivotTable.criteria object
}
\description{
PivotTable.criteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An optional mapping of filters per source column offset.The filters will be applied before aggregating data into the pivot table.The map's key is the column offset of the source range that you want tofilter, and the value is the criteria for that column.For example, if the source was `C10:E15`, a key of `0` will have the filterfor column `C`, whereas the key `1` is for column `D`.
}
\seealso{
Other PivotTable functions: \code{\link{PivotTable}}
}

