% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotTable}
\alias{PivotTable}
\title{PivotTable Object}
\usage{
PivotTable(PivotTable.criteria = NULL, valueLayout = NULL, columns = NULL,
  source = NULL, rows = NULL, values = NULL, criteria = NULL)
}
\arguments{
\item{PivotTable.criteria}{The \link{PivotTable.criteria} object or list of objects}

\item{valueLayout}{Whether values should be listed horizontally (as columns)}

\item{columns}{Each column grouping in the pivot table}

\item{source}{The range the pivot table is reading data from}

\item{rows}{Each row grouping in the pivot table}

\item{values}{A list of values to include in the pivot table}

\item{criteria}{An optional mapping of filters per source column offset}
}
\value{
PivotTable object
}
\description{
PivotTable Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A pivot table.
}
\seealso{
Other PivotTable functions: \code{\link{PivotTable.criteria}}
}

