% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotValue}
\alias{PivotValue}
\title{PivotValue Object}
\usage{
PivotValue(formula = NULL, sourceColumnOffset = NULL,
  summarizeFunction = NULL, name = NULL)
}
\arguments{
\item{formula}{A custom formula to calculate the value}

\item{sourceColumnOffset}{The column offset of the source range that this value reads from}

\item{summarizeFunction}{A function to summarize the value}

\item{name}{A name to use for the value}
}
\value{
PivotValue object
}
\description{
PivotValue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The definition of how a value in a pivot table should be calculated.
}

