% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ProtectedRange}
\alias{ProtectedRange}
\title{ProtectedRange Object}
\usage{
ProtectedRange(unprotectedRanges = NULL, description = NULL,
  namedRangeId = NULL, requestingUserCanEdit = NULL, editors = NULL,
  protectedRangeId = NULL, warningOnly = NULL, range = NULL)
}
\arguments{
\item{unprotectedRanges}{The list of unprotected ranges within a protected sheet}

\item{description}{The description of this protected range}

\item{namedRangeId}{The named range this protected range is backed by, if any}

\item{requestingUserCanEdit}{True if the user who requested this protected range can edit the}

\item{editors}{The users and groups with edit access to the protected range}

\item{protectedRangeId}{The ID of the protected range}

\item{warningOnly}{True if this protected range will show a warning when editing}

\item{range}{The range that is being protected}
}
\value{
ProtectedRange object
}
\description{
ProtectedRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A protected range.
}

