% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{RepeatCellRequest}
\alias{RepeatCellRequest}
\title{RepeatCellRequest Object}
\usage{
RepeatCellRequest(range = NULL, fields = NULL, cell = NULL)
}
\arguments{
\item{range}{The range to repeat the cell in}

\item{fields}{The fields that should be updated}

\item{cell}{The data to write}
}
\value{
RepeatCellRequest object
}
\description{
RepeatCellRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates all cells in the range to the values in the given Cell object.Only the fields listed in the fields field are updated; others areunchanged.If writing a cell with a formula, the formula's ranges will automaticallyincrement for each field in the range.For example, if writing a cell with formula `=A1` into range B2:C4,B2 would be `=A1`, B3 would be `=A2`, B4 would be `=A3`,C2 would be `=B1`, C3 would be `=B2`, C4 would be `=B3`.To keep the formula's ranges static, use the `$` indicator.For example, use the formula `=$A$1` to prevent both the row and thecolumn from incrementing.
}
