% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{RowData}
\alias{RowData}
\title{RowData Object}
\usage{
RowData(values = NULL)
}
\arguments{
\item{values}{The values in the row, one per column}
}
\value{
RowData object
}
\description{
RowData Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data about each cell in a row.
}
