% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SetBasicFilterRequest}
\alias{SetBasicFilterRequest}
\title{SetBasicFilterRequest Object}
\usage{
SetBasicFilterRequest(filter = NULL)
}
\arguments{
\item{filter}{The filter to set}
}
\value{
SetBasicFilterRequest object
}
\description{
SetBasicFilterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sets the basic filter associated with a sheet.
}
