% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SetDataValidationRequest}
\alias{SetDataValidationRequest}
\title{SetDataValidationRequest Object}
\usage{
SetDataValidationRequest(rule = NULL, range = NULL)
}
\arguments{
\item{rule}{The data validation rule to set on each cell in the range,}

\item{range}{The range the data validation rule should apply to}
}
\value{
SetDataValidationRequest object
}
\description{
SetDataValidationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sets a data validation rule to every cell in the range.To clear validation in a range, call this with no rule specified.
}
