% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Sheet}
\alias{Sheet}
\title{Sheet Object}
\usage{
Sheet(bandedRanges = NULL, charts = NULL, properties = NULL,
  filterViews = NULL, conditionalFormats = NULL, protectedRanges = NULL,
  basicFilter = NULL, merges = NULL, data = NULL)
}
\arguments{
\item{bandedRanges}{The banded (i}

\item{charts}{The specifications of every chart on this sheet}

\item{properties}{The properties of the sheet}

\item{filterViews}{The filter views in this sheet}

\item{conditionalFormats}{The conditional format rules in this sheet}

\item{protectedRanges}{The protected ranges in this sheet}

\item{basicFilter}{The filter on this sheet, if any}

\item{merges}{The ranges that are merged together}

\item{data}{Data in the grid, if this is a grid sheet}
}
\value{
Sheet object
}
\description{
Sheet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A sheet in a spreadsheet.
}
