% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SortRangeRequest}
\alias{SortRangeRequest}
\title{SortRangeRequest Object}
\usage{
SortRangeRequest(range = NULL, sortSpecs = NULL)
}
\arguments{
\item{range}{The range to sort}

\item{sortSpecs}{The sort order per column}
}
\value{
SortRangeRequest object
}
\description{
SortRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Sorts data in rows based on a sort order per column.
}
