% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SourceAndDestination}
\alias{SourceAndDestination}
\title{SourceAndDestination Object}
\usage{
SourceAndDestination(source = NULL, fillLength = NULL, dimension = NULL)
}
\arguments{
\item{source}{The location of the data to use as the source of the autofill}

\item{fillLength}{The number of rows or columns that data should be filled into}

\item{dimension}{The dimension that data should be filled into}
}
\value{
SourceAndDestination object
}
\description{
SourceAndDestination Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A combination of a source range and how to extend that source.
}

