% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadbatchUpdate}
\alias{spreadbatchUpdate}
\title{Applies one or more updates to the spreadsheet.Each request is validated beforebeing applied. If any request is not valid then the entire request willfail and nothing will be applied.Some requests have replies togive you some information about howthey are applied. The replies will mirror the requests.  For example,if you applied 4 updates and the 3rd one had a reply, then theresponse will have 2 empty replies, the actual reply, and another emptyreply, in that order.Due to the collaborative nature of spreadsheets, it is not guaranteed thatthe spreadsheet will reflect exactly your changes after this completes,however it is guaranteed that the updates in the request will beapplied together atomically. Your changes may be altered with respect tocollaborator changes. If there are no collaborators, the spreadsheetshould reflect your changes.}
\usage{
spreadbatchUpdate(BatchUpdateSpreadsheetRequest, spreadsheetId)
}
\arguments{
\item{BatchUpdateSpreadsheetRequest}{The \link{BatchUpdateSpreadsheetRequest} object to pass to this method}

\item{spreadsheetId}{The spreadsheet to apply the updates to}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/spreadsheets
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/spreadsheets)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}

Other BatchUpdateSpreadsheetRequest functions: \code{\link{BatchUpdateSpreadsheetRequest}}
}

