% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_functions.R
\name{spreadget}
\alias{spreadget}
\title{Returns the spreadsheet at the given ID.The caller must specify the spreadsheet ID.By default, data within grids will not be returned.You can include grid data one of two ways:* Specify a field mask listing your desired fields using the `fields` URLparameter in HTTP* Set the includeGridDataURL parameter to true.  If a field mask is set, the `includeGridData`parameter is ignoredFor large spreadsheets, it is recommended to retrieve only the specificfields of the spreadsheet that you want.To retrieve only subsets of the spreadsheet, use theranges URL parameter.Multiple ranges can be specified.  Limiting the range willreturn only the portions of the spreadsheet that intersect the requestedranges. Ranges are specified using A1 notation.}
\usage{
spreadget(spreadsheetId, ranges = NULL, includeGridData = NULL)
}
\arguments{
\item{spreadsheetId}{The spreadsheet to request}

\item{ranges}{The ranges to retrieve from the spreadsheet}

\item{includeGridData}{True if grid data should be returned}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.readonly
\item https://www.googleapis.com/auth/spreadsheets
\item https://www.googleapis.com/auth/spreadsheets.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.readonly, https://www.googleapis.com/auth/spreadsheets, https://www.googleapis.com/auth/spreadsheets.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/sheets/}{Google Documentation}
}
