% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Spreadsheet}
\alias{Spreadsheet}
\title{Spreadsheet Object}
\usage{
Spreadsheet(properties = NULL, spreadsheetId = NULL, sheets = NULL,
  namedRanges = NULL, spreadsheetUrl = NULL)
}
\arguments{
\item{properties}{Overall properties of a spreadsheet}

\item{spreadsheetId}{The ID of the spreadsheet}

\item{sheets}{The sheets that are part of a spreadsheet}

\item{namedRanges}{The named ranges defined in a spreadsheet}

\item{spreadsheetUrl}{The url of the spreadsheet}
}
\value{
Spreadsheet object
}
\description{
Spreadsheet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Resource that represents a spreadsheet.
}
\seealso{
Other Spreadsheet functions: \code{\link{spreadcreate}}
}
