% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{TextFormatRun}
\alias{TextFormatRun}
\title{TextFormatRun Object}
\usage{
TextFormatRun(startIndex = NULL, format = NULL)
}
\arguments{
\item{startIndex}{The character index where this run starts}

\item{format}{The format of this run}
}
\value{
TextFormatRun object
}
\description{
TextFormatRun Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A run of a text format. The format of this run continues until the startindex of the next run.When updating, all fields must be set.
}
