% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UnmergeCellsRequest}
\alias{UnmergeCellsRequest}
\title{UnmergeCellsRequest Object}
\usage{
UnmergeCellsRequest(range = NULL)
}
\arguments{
\item{range}{The range within which all cells should be unmerged}
}
\value{
UnmergeCellsRequest object
}
\description{
UnmergeCellsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Unmerges cells in the given range.
}
