% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateBandingRequest}
\alias{UpdateBandingRequest}
\title{UpdateBandingRequest Object}
\usage{
UpdateBandingRequest(fields = NULL, bandedRange = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{bandedRange}{The banded range to update with the new properties}
}
\value{
UpdateBandingRequest object
}
\description{
UpdateBandingRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates properties of the supplied banded range.
}
