% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateBordersRequest}
\alias{UpdateBordersRequest}
\title{UpdateBordersRequest Object}
\usage{
UpdateBordersRequest(right = NULL, innerVertical = NULL, top = NULL,
  innerHorizontal = NULL, range = NULL, bottom = NULL, left = NULL)
}
\arguments{
\item{right}{The border to put at the right of the range}

\item{innerVertical}{The vertical border to put within the range}

\item{top}{The border to put at the top of the range}

\item{innerHorizontal}{The horizontal border to put within the range}

\item{range}{The range whose borders should be updated}

\item{bottom}{The border to put at the bottom of the range}

\item{left}{The border to put at the left of the range}
}
\value{
UpdateBordersRequest object
}
\description{
UpdateBordersRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the borders of a range.If a field is not set in the request, that means the border remains as-is.For example, with two subsequent UpdateBordersRequest: 1. range: A1:A5 `{ top: RED, bottom: WHITE }` 2. range: A1:A5 `{ left: BLUE }`That would result in A1:A5 having a borders of`{ top: RED, bottom: WHITE, left: BLUE }`.If you want to clear a border, explicitly set the style toNONE.
}

