% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateChartSpecRequest}
\alias{UpdateChartSpecRequest}
\title{UpdateChartSpecRequest Object}
\usage{
UpdateChartSpecRequest(chartId = NULL, spec = NULL)
}
\arguments{
\item{chartId}{The ID of the chart to update}

\item{spec}{The specification to apply to the chart}
}
\value{
UpdateChartSpecRequest object
}
\description{
UpdateChartSpecRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates a chart's specifications.(This does not move or resize a chart. To move or resize a chart, use UpdateEmbeddedObjectPositionRequest.)
}
