% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateConditionalFormatRuleResponse}
\alias{UpdateConditionalFormatRuleResponse}
\title{UpdateConditionalFormatRuleResponse Object}
\usage{
UpdateConditionalFormatRuleResponse(oldIndex = NULL, newRule = NULL,
  oldRule = NULL, newIndex = NULL)
}
\arguments{
\item{oldIndex}{The old index of the rule}

\item{newRule}{The new rule that replaced the old rule (if replacing),}

\item{oldRule}{The old (deleted) rule}

\item{newIndex}{The index of the new rule}
}
\value{
UpdateConditionalFormatRuleResponse object
}
\description{
UpdateConditionalFormatRuleResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of updating a conditional format rule.
}
