% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateEmbeddedObjectPositionResponse}
\alias{UpdateEmbeddedObjectPositionResponse}
\title{UpdateEmbeddedObjectPositionResponse Object}
\usage{
UpdateEmbeddedObjectPositionResponse(position = NULL)
}
\arguments{
\item{position}{The new position of the embedded object}
}
\value{
UpdateEmbeddedObjectPositionResponse object
}
\description{
UpdateEmbeddedObjectPositionResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of updating an embedded object's position.
}
