% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateFilterViewRequest}
\alias{UpdateFilterViewRequest}
\title{UpdateFilterViewRequest Object}
\usage{
UpdateFilterViewRequest(filter = NULL, fields = NULL)
}
\arguments{
\item{filter}{The new properties of the filter view}

\item{fields}{The fields that should be updated}
}
\value{
UpdateFilterViewRequest object
}
\description{
UpdateFilterViewRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates properties of the filter view.
}

