% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateProtectedRangeRequest}
\alias{UpdateProtectedRangeRequest}
\title{UpdateProtectedRangeRequest Object}
\usage{
UpdateProtectedRangeRequest(protectedRange = NULL, fields = NULL)
}
\arguments{
\item{protectedRange}{The protected range to update with the new properties}

\item{fields}{The fields that should be updated}
}
\value{
UpdateProtectedRangeRequest object
}
\description{
UpdateProtectedRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates an existing protected range with the specifiedprotectedRangeId.
}
