% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateSheetPropertiesRequest}
\alias{UpdateSheetPropertiesRequest}
\title{UpdateSheetPropertiesRequest Object}
\usage{
UpdateSheetPropertiesRequest(properties = NULL, fields = NULL)
}
\arguments{
\item{properties}{The properties to update}

\item{fields}{The fields that should be updated}
}
\value{
UpdateSheetPropertiesRequest object
}
\description{
UpdateSheetPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates properties of the sheet with the specifiedsheetId.
}
