% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateValuesResponse}
\alias{UpdateValuesResponse}
\title{UpdateValuesResponse Object}
\usage{
UpdateValuesResponse(updatedRange = NULL, updatedColumns = NULL,
  spreadsheetId = NULL, updatedRows = NULL, updatedCells = NULL)
}
\arguments{
\item{updatedRange}{The range (in A1 notation) that updates were applied to}

\item{updatedColumns}{The number of columns where at least one cell in the column was updated}

\item{spreadsheetId}{The spreadsheet the updates were applied to}

\item{updatedRows}{The number of rows where at least one cell in the row was updated}

\item{updatedCells}{The number of cells updated}
}
\value{
UpdateValuesResponse object
}
\description{
UpdateValuesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response when updating a range of values in a spreadsheet.
}

