% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ValueRange}
\alias{ValueRange}
\title{ValueRange Object}
\usage{
ValueRange(majorDimension = NULL, values = NULL, range = NULL)
}
\arguments{
\item{majorDimension}{The major dimension of the values}

\item{values}{The data that was read or to be written}

\item{range}{The range the values cover, in A1 notation}
}
\value{
ValueRange object
}
\description{
ValueRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Data within a range of the spreadsheet.
}
