% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{AddNamedRangeResponse}
\alias{AddNamedRangeResponse}
\title{AddNamedRangeResponse Object}
\usage{
AddNamedRangeResponse(namedRange = NULL)
}
\arguments{
\item{namedRange}{The named range to add}
}
\value{
AddNamedRangeResponse object
}
\description{
AddNamedRangeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of adding a named range.
}
