% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BasicChartSpec}
\alias{BasicChartSpec}
\title{BasicChartSpec Object}
\usage{
BasicChartSpec(headerCount = NULL, axis = NULL, chartType = NULL,
  series = NULL, legendPosition = NULL, domains = NULL)
}
\arguments{
\item{headerCount}{The number of rows or columns in the data that are 'headers'}

\item{axis}{The axis on the chart}

\item{chartType}{The type of the chart}

\item{series}{The data this chart is visualizing}

\item{legendPosition}{The position of the chart legend}

\item{domains}{The domain of data this is charting}
}
\value{
BasicChartSpec object
}
\description{
BasicChartSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The specification for a basic chart.  See BasicChartType for the listof charts this supports.
}
