% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{BooleanRule}
\alias{BooleanRule}
\title{BooleanRule Object}
\usage{
BooleanRule(format = NULL, condition = NULL)
}
\arguments{
\item{format}{The format to apply}

\item{condition}{The condition of the rule}
}
\value{
BooleanRule object
}
\description{
BooleanRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rule that may or may not match, depending on the condition.
}
