% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CellFormat}
\alias{CellFormat}
\title{CellFormat Object}
\usage{
CellFormat(horizontalAlignment = NULL, hyperlinkDisplayType = NULL,
  borders = NULL, textDirection = NULL, textFormat = NULL,
  padding = NULL, numberFormat = NULL, wrapStrategy = NULL,
  backgroundColor = NULL, verticalAlignment = NULL)
}
\arguments{
\item{horizontalAlignment}{The horizontal alignment of the value in the cell}

\item{hyperlinkDisplayType}{How a hyperlink, if it exists, should be displayed in the cell}

\item{borders}{The borders of the cell}

\item{textDirection}{The direction of the text in the cell}

\item{textFormat}{The format of the text in the cell (unless overridden by a format run)}

\item{padding}{The padding of the cell}

\item{numberFormat}{A format describing how number values should be represented to the user}

\item{wrapStrategy}{The wrap strategy for the value in the cell}

\item{backgroundColor}{The background color of the cell}

\item{verticalAlignment}{The vertical alignment of the value in the cell}
}
\value{
CellFormat object
}
\description{
CellFormat Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The format of a cell.
}

