% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{CutPasteRequest}
\alias{CutPasteRequest}
\title{CutPasteRequest Object}
\usage{
CutPasteRequest(pasteType = NULL, source = NULL, destination = NULL)
}
\arguments{
\item{pasteType}{What kind of data to paste}

\item{source}{The source data to cut}

\item{destination}{The top-left coordinate where the data should be pasted}
}
\value{
CutPasteRequest object
}
\description{
CutPasteRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Moves data from the source to the destination.
}

