% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{DeleteConditionalFormatRuleRequest}
\alias{DeleteConditionalFormatRuleRequest}
\title{DeleteConditionalFormatRuleRequest Object}
\usage{
DeleteConditionalFormatRuleRequest(index = NULL, sheetId = NULL)
}
\arguments{
\item{index}{The zero-based index of the rule to be deleted}

\item{sheetId}{The sheet the rule is being deleted from}
}
\value{
DeleteConditionalFormatRuleRequest object
}
\description{
DeleteConditionalFormatRuleRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes a conditional format rule at the given index.All subsequent rules' indexes are decremented.
}
