% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Editors}
\alias{Editors}
\title{Editors Object}
\usage{
Editors(groups = NULL, domainUsersCanEdit = NULL, users = NULL)
}
\arguments{
\item{groups}{The email addresses of groups with edit access to the protected range}

\item{domainUsersCanEdit}{True if anyone in the document's domain has edit access to the protected}

\item{users}{The email addresses of users with edit access to the protected range}
}
\value{
Editors object
}
\description{
Editors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The editors of a protected range.
}
