% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{EmbeddedObjectPosition}
\alias{EmbeddedObjectPosition}
\title{EmbeddedObjectPosition Object}
\usage{
EmbeddedObjectPosition(sheetId = NULL, overlayPosition = NULL,
  newSheet = NULL)
}
\arguments{
\item{sheetId}{The sheet this is on}

\item{overlayPosition}{The position at which the object is overlaid on top of a grid}

\item{newSheet}{If true, the embedded object will be put on a new sheet whose ID}
}
\value{
EmbeddedObjectPosition object
}
\description{
EmbeddedObjectPosition Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The position of an embedded object such as a chart.
}

