% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{FilterCriteria}
\alias{FilterCriteria}
\title{FilterCriteria Object}
\usage{
FilterCriteria(hiddenValues = NULL, condition = NULL)
}
\arguments{
\item{hiddenValues}{Values that should be hidden}

\item{condition}{A condition that must be true for values to be shown}
}
\value{
FilterCriteria object
}
\description{
FilterCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Criteria for showing/hiding rows in a filter or filter view.
}
