% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{FindReplaceResponse}
\alias{FindReplaceResponse}
\title{FindReplaceResponse Object}
\usage{
FindReplaceResponse(sheetsChanged = NULL, formulasChanged = NULL,
  valuesChanged = NULL, occurrencesChanged = NULL, rowsChanged = NULL)
}
\arguments{
\item{sheetsChanged}{The number of sheets changed}

\item{formulasChanged}{The number of formula cells changed}

\item{valuesChanged}{The number of non-formula cells changed}

\item{occurrencesChanged}{The number of occurrences (possibly multiple within a cell) changed}

\item{rowsChanged}{The number of rows changed}
}
\value{
FindReplaceResponse object
}
\description{
FindReplaceResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of the find/replace.
}
