% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{GradientRule}
\alias{GradientRule}
\title{GradientRule Object}
\usage{
GradientRule(midpoint = NULL, minpoint = NULL, maxpoint = NULL)
}
\arguments{
\item{midpoint}{An optional midway interpolation point}

\item{minpoint}{The starting interpolation point}

\item{maxpoint}{The final interpolation point}
}
\value{
GradientRule object
}
\description{
GradientRule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A rule that applies a gradient color scale format, based onthe interpolation points listed. The format of a cell will varybased on its contents as compared to the values of the interpolationpoints.
}
