% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{GridCoordinate}
\alias{GridCoordinate}
\title{GridCoordinate Object}
\usage{
GridCoordinate(rowIndex = NULL, columnIndex = NULL, sheetId = NULL)
}
\arguments{
\item{rowIndex}{The row index of the coordinate}

\item{columnIndex}{The column index of the coordinate}

\item{sheetId}{The sheet this coordinate is on}
}
\value{
GridCoordinate object
}
\description{
GridCoordinate Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A coordinate in a sheet.All indexes are zero-based.
}
