% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{InterpolationPoint}
\alias{InterpolationPoint}
\title{InterpolationPoint Object}
\usage{
InterpolationPoint(value = NULL, color = NULL, type = NULL)
}
\arguments{
\item{value}{The value this interpolation point uses}

\item{color}{The color this interpolation point should use}

\item{type}{How the value should be interpreted}
}
\value{
InterpolationPoint object
}
\description{
InterpolationPoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single interpolation point on a gradient conditional format.These pin the gradient color scale according to the color,type and value chosen.
}

