% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotGroup}
\alias{PivotGroup}
\title{PivotGroup Object}
\usage{
PivotGroup(sourceColumnOffset = NULL, showTotals = NULL,
  valueMetadata = NULL, sortOrder = NULL, valueBucket = NULL)
}
\arguments{
\item{sourceColumnOffset}{The column offset of the source range that this grouping is based on}

\item{showTotals}{True if the pivot table should include the totals for this grouping}

\item{valueMetadata}{Metadata about values in the grouping}

\item{sortOrder}{The order the values in this group should be sorted}

\item{valueBucket}{The bucket of the opposite pivot group to sort by}
}
\value{
PivotGroup object
}
\description{
PivotGroup Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single grouping (either row or column) in a pivot table.
}
