% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{PivotGroupSortValueBucket}
\alias{PivotGroupSortValueBucket}
\title{PivotGroupSortValueBucket Object}
\usage{
PivotGroupSortValueBucket(valuesIndex = NULL, buckets = NULL)
}
\arguments{
\item{valuesIndex}{The offset in the PivotTable}

\item{buckets}{Determines the bucket from which values are chosen to sort}
}
\value{
PivotGroupSortValueBucket object
}
\description{
PivotGroupSortValueBucket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about which values in a pivot group should be used for sorting.
}
