% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{ProtectedRange}
\alias{ProtectedRange}
\title{ProtectedRange Object}
\usage{
ProtectedRange(description = NULL, unprotectedRanges = NULL,
  namedRangeId = NULL, protectedRangeId = NULL, warningOnly = NULL,
  requestingUserCanEdit = NULL, range = NULL, editors = NULL)
}
\arguments{
\item{description}{The description of this protected range}

\item{unprotectedRanges}{The list of unprotected ranges within a protected sheet}

\item{namedRangeId}{The named range this protected range is backed by, if any}

\item{protectedRangeId}{The ID of the protected range}

\item{warningOnly}{True if this protected range will show a warning when editing}

\item{requestingUserCanEdit}{True if the user who requested this protected range can edit the}

\item{range}{The range that is being protected}

\item{editors}{The users and groups with edit access to the protected range}
}
\value{
ProtectedRange object
}
\description{
ProtectedRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A protected range.
}
