% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{Request}
\alias{Request}
\title{Request Object}
\usage{
Request(updateEmbeddedObjectPosition = NULL, deleteNamedRange = NULL,
  updateNamedRange = NULL, addFilterView = NULL,
  updateSpreadsheetProperties = NULL, appendDimension = NULL,
  unmergeCells = NULL, updateProtectedRange = NULL,
  deleteFilterView = NULL, clearBasicFilter = NULL, sortRange = NULL,
  repeatCell = NULL, setDataValidation = NULL, updateCells = NULL,
  addSheet = NULL, updateFilterView = NULL, updateSheetProperties = NULL,
  updateDimensionProperties = NULL, deleteSheet = NULL,
  findReplace = NULL, addProtectedRange = NULL,
  deleteProtectedRange = NULL, updateConditionalFormatRule = NULL,
  setBasicFilter = NULL, mergeCells = NULL, addChart = NULL,
  deleteConditionalFormatRule = NULL, updateChartSpec = NULL,
  deleteDimension = NULL, deleteEmbeddedObject = NULL, pasteData = NULL,
  addConditionalFormatRule = NULL, updateBorders = NULL,
  autoResizeDimensions = NULL, duplicateSheet = NULL,
  duplicateFilterView = NULL, cutPaste = NULL, appendCells = NULL,
  addNamedRange = NULL, autoFill = NULL, moveDimension = NULL,
  textToColumns = NULL, insertDimension = NULL, copyPaste = NULL)
}
\arguments{
\item{updateEmbeddedObjectPosition}{Updates an embedded object's (e}

\item{deleteNamedRange}{Deletes a named range}

\item{updateNamedRange}{Updates a named range}

\item{addFilterView}{Adds a filter view}

\item{updateSpreadsheetProperties}{Updates the spreadsheet's properties}

\item{appendDimension}{Appends dimensions to the end of a sheet}

\item{unmergeCells}{Unmerges merged cells}

\item{updateProtectedRange}{Updates a protected range}

\item{deleteFilterView}{Deletes a filter view from a sheet}

\item{clearBasicFilter}{Clears the basic filter on a sheet}

\item{sortRange}{Sorts data in a range}

\item{repeatCell}{Repeats a single cell across a range}

\item{setDataValidation}{Sets data validation for one or more cells}

\item{updateCells}{Updates many cells at once}

\item{addSheet}{Adds a sheet}

\item{updateFilterView}{Updates the properties of a filter view}

\item{updateSheetProperties}{Updates a sheet's properties}

\item{updateDimensionProperties}{Updates dimensions' properties}

\item{deleteSheet}{Deletes a sheet}

\item{findReplace}{Finds and replaces occurrences of some text with other text}

\item{addProtectedRange}{Adds a protected range}

\item{deleteProtectedRange}{Deletes a protected range}

\item{updateConditionalFormatRule}{Updates an existing conditional format rule}

\item{setBasicFilter}{Sets the basic filter on a sheet}

\item{mergeCells}{Merges cells together}

\item{addChart}{Adds a chart}

\item{deleteConditionalFormatRule}{Deletes an existing conditional format rule}

\item{updateChartSpec}{Updates a chart's specifications}

\item{deleteDimension}{Deletes rows or columns in a sheet}

\item{deleteEmbeddedObject}{Deletes an embedded object (e}

\item{pasteData}{Pastes data (HTML or delimited) into a sheet}

\item{addConditionalFormatRule}{Adds a new conditional format rule}

\item{updateBorders}{Updates the borders in a range of cells}

\item{autoResizeDimensions}{Automatically resizes one or more dimensions based on the contents}

\item{duplicateSheet}{Duplicates a sheet}

\item{duplicateFilterView}{Duplicates a filter view}

\item{cutPaste}{Cuts data from one area and pastes it to another}

\item{appendCells}{Appends cells after the last row with data in a sheet}

\item{addNamedRange}{Adds a named range}

\item{autoFill}{Automatically fills in more data based on existing data}

\item{moveDimension}{Moves rows or columns to another location in a sheet}

\item{textToColumns}{Converts a column of text into many columns of text}

\item{insertDimension}{Inserts new rows or columns in a sheet}

\item{copyPaste}{Copies data from one area and pastes it to another}
}
\value{
Request object
}
\description{
Request Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single kind of update to apply to a spreadsheet.
}

