% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SheetProperties}
\alias{SheetProperties}
\title{SheetProperties Object}
\usage{
SheetProperties(title = NULL, index = NULL, hidden = NULL,
  gridProperties = NULL, sheetId = NULL, rightToLeft = NULL,
  tabColor = NULL, sheetType = NULL)
}
\arguments{
\item{title}{The name of the sheet}

\item{index}{The index of the sheet within the spreadsheet}

\item{hidden}{True if the sheet is hidden in the UI, false if it's visible}

\item{gridProperties}{Additional properties of the sheet if this sheet is a grid}

\item{sheetId}{The ID of the sheet}

\item{rightToLeft}{True if the sheet is an RTL sheet instead of an LTR sheet}

\item{tabColor}{The color of the tab in the UI}

\item{sheetType}{The type of sheet}
}
\value{
SheetProperties object
}
\description{
SheetProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of a sheet.
}

