% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{SpreadsheetProperties}
\alias{SpreadsheetProperties}
\title{SpreadsheetProperties Object}
\usage{
SpreadsheetProperties(iterativeCalculationSettings = NULL,
  defaultFormat = NULL, autoRecalc = NULL, title = NULL,
  timeZone = NULL, locale = NULL)
}
\arguments{
\item{iterativeCalculationSettings}{Determines whether and how circular references are resolved with iterative}

\item{defaultFormat}{The default format of all cells in the spreadsheet}

\item{autoRecalc}{The amount of time to wait before volatile functions are recalculated}

\item{title}{The title of the spreadsheet}

\item{timeZone}{The time zone of the spreadsheet, in CLDR format such as}

\item{locale}{The locale of the spreadsheet in one of the following formats:}
}
\value{
SpreadsheetProperties object
}
\description{
SpreadsheetProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of a spreadsheet.
}
