% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{TextToColumnsRequest}
\alias{TextToColumnsRequest}
\title{TextToColumnsRequest Object}
\usage{
TextToColumnsRequest(delimiter = NULL, source = NULL,
  delimiterType = NULL)
}
\arguments{
\item{delimiter}{The delimiter to use}

\item{source}{The source data range}

\item{delimiterType}{The delimiter type to use}
}
\value{
TextToColumnsRequest object
}
\description{
TextToColumnsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Splits a column of text into multiple columns,based on a delimiter in each cell.
}
