% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateDimensionPropertiesRequest}
\alias{UpdateDimensionPropertiesRequest}
\title{UpdateDimensionPropertiesRequest Object}
\usage{
UpdateDimensionPropertiesRequest(fields = NULL, range = NULL,
  properties = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{range}{The rows or columns to update}

\item{properties}{Properties to update}
}
\value{
UpdateDimensionPropertiesRequest object
}
\description{
UpdateDimensionPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates properties of dimensions within the specified range.
}

