% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateEmbeddedObjectPositionRequest}
\alias{UpdateEmbeddedObjectPositionRequest}
\title{UpdateEmbeddedObjectPositionRequest Object}
\usage{
UpdateEmbeddedObjectPositionRequest(newPosition = NULL, objectId = NULL,
  fields = NULL)
}
\arguments{
\item{newPosition}{An explicit position to move the embedded object to}

\item{objectId}{The ID of the object to moved}

\item{fields}{The fields of OverlayPosition}
}
\value{
UpdateEmbeddedObjectPositionRequest object
}
\description{
UpdateEmbeddedObjectPositionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update an embedded object's position (such as a moving or resizing achart or image).
}

