% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateNamedRangeRequest}
\alias{UpdateNamedRangeRequest}
\title{UpdateNamedRangeRequest Object}
\usage{
UpdateNamedRangeRequest(namedRange = NULL, fields = NULL)
}
\arguments{
\item{namedRange}{The named range to update with the new properties}

\item{fields}{The fields that should be updated}
}
\value{
UpdateNamedRangeRequest object
}
\description{
UpdateNamedRangeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates properties of the named range with the specifiednamedRangeId.
}
