% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheets_objects.R
\name{UpdateSpreadsheetPropertiesRequest}
\alias{UpdateSpreadsheetPropertiesRequest}
\title{UpdateSpreadsheetPropertiesRequest Object}
\usage{
UpdateSpreadsheetPropertiesRequest(fields = NULL, properties = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{properties}{The properties to update}
}
\value{
UpdateSpreadsheetPropertiesRequest object
}
\description{
UpdateSpreadsheetPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates properties of a spreadsheet.
}

